Modified 'feature-viewer.ts' and 'fillsvg.ts'. To support user defined IDs.
Found a problem with selecting a feature in a certain case where one can
have two features on same position and same length. For example user can
have 2 SNPs on same position but diffent outcome. In usual behaviour of
the library one gets two lines with the features under eachother.

The problem arises when trying to select the second feature. One clicks
on it but is selected the first one. That happens because in 'fillsvg.ts'
the features are automatically given IDs in the format: 'f_FEAT_x_y', hence
the two features recieve same id.

After the changes if the user has given unique ids in proper format
which can be used by selectors these IDs are being used, else the ID is being
assigned as before.

Note: Changes are valid only for features with type: rect and unique.